+++ test/unit/chat_hide_test.dart
@@ -314,7 +314,7 @@ void main() async {
     ChatService chatService = Get.put(ChatService(chatRepository, authService));
 
     expect(
-      () async => await chatService.hideChat(
+      chatService.hideChat(
         const ChatId('0d72d245-8425-467a-9ebd-082d4f47850b'),
       ),
       throwsA(isA<HideChatException>()),