+++ CHANGELOG.md
@@ -16,6 +16,8 @@ All user visible changes to this project will be documented in this file. This p
 - UI:
     - Chat page:
         - Call buttons position setting and adjusting. ([#750], [#718])
+    - Chats tab:
+        - Monolog searching. ([#745], [#582])
 
 ### Fixed
 
@@ -27,8 +29,10 @@ All user visible changes to this project will be documented in this file. This p
     - User page:
         - Downloaded avatar missing its extension on desktop. ([#756, #726])
 
+[#582]: /../../issues/582
 [#718]: /../../issues/718
 [#726]: /../../issues/726
+[#745]: /../../pull/745
 [#750]: /../../pull/750
 [#756]: /../../pull/756
 [#758]: /../../issues/758